+++ src/components/CloseButton.tsx
@@ -6,26 +6,15 @@ interface Props {
   onClick: () => void;
   size?: CSSProperties["width"];
   variant?: "light" | "dark";
-  className?: string;
 }
-export function CloseButton({
-  onClick,
-  size = 14,
-  variant = "light",
-  className,
-}: Props) {
+export function CloseButton({ onClick, size = 14, variant = "light" }: Props) {
   const style = {
     "--size": typeof size === "number" ? `${size}px` : size,
     "--fill": variant === "light" ? white : darkText,
   } as CSSProperties;
 
   return (
-    <Button
-      onClick={onClick}
-      style={style}
-      aria-label="Close"
-      className={className}
-    >
+    <Button onClick={onClick} style={style} aria-label="Close">
       <CloseIcon />
     </Button>
   );
