+++ src/components/Panel/OoTypeIcon.tsx
@@ -1,8 +1,8 @@
 import { IconWrapper } from "@/components";
+import { InfoIconText, InfoIconWrapper } from "@/components/style";
 import { OracleType } from "@/types";
 import OptimisticOracle from "public/assets/icons/optimistic-oracle.svg";
 import styled from "styled-components";
-import { InfoIconText, InfoIconWrapper } from "./style";
 
 /**
  * Displays an icon for the given Optimistic Oracle type.
