--- src/components/Panel/style.tsx
@@ -1,17 +1,0 @@
-import styled from "styled-components";
-
-export const InfoIconWrapper = styled.div`
-  height: 35px;
-  width: max-content;
-  display: flex;
-  align-items: center;
-  gap: 10px;
-  padding-inline: 10px;
-  padding-block: 8px;
-  border: 1px solid var(--grey-100);
-  border-radius: 5px;
-`;
-
-export const InfoIconText = styled.p`
-  font: var(--body-xs);
-`;
