+++ src/components/index.ts
@@ -1,20 +1,19 @@
 export { Button } from "./Button";
-export { Checkbox } from "./Checkbox/Checkbox";
-export * from "./Checkbox/style";
+export { Checkbox } from "./Checkbox";
+export { CheckboxDropdown } from "./CheckboxDropdown";
 export { CloseButton } from "./CloseButton";
 export { ConnectButton } from "./ConnectButton";
 export { DecimalInput } from "./DecimalInput";
-export { CheckboxDropdown } from "./Dropdown/CheckboxDropdown";
-export { RadioDropdown } from "./Dropdown/RadioDropdown";
 export { ErrorBanner } from "./ErrorBanner";
 export { ErrorMessage } from "./ErrorMessage";
-export { Filters } from "./Filters/Filters";
+export { Filters } from "./Filters";
 export { GlobalStyle } from "./GlobalStyle";
 export { Header } from "./Header/Header";
 export { IconWrapper } from "./IconWrapper";
 export { Layout } from "./Layout";
+export { MobileFilters } from "./MobileFilters";
+export { MobileMenu } from "./MobileMenu";
 export { Pagination } from "./Pagination";
-export { MobileFilters } from "./Panel/MobileFilters";
-export { MobileMenu } from "./Panel/MobileMenu";
-export { Panel } from "./Panel/Panel";
+export { Panel } from "./Panel";
+export { RadioDropdown } from "./RadioDropdown";
 export { Table } from "./Table/Table";
